+++ packages/client/hmi-client/src/utils/petri-net-validator.ts
